+++ src/detect.ts
@@ -52,9 +52,13 @@ export const isDevProxyRunning = async (devProxyExe: string): Promise<boolean> =
         return processId.trim() !== '';
     };
     if (platform === 'darwin') {
-        const processId = await executeCommand(`$SHELL -c "ps -ef | grep ${devProxyExe} | grep -v grep | awk \'{print $2}\'"`);
+        const processId = await executeCommand(`$SHELL -c "ps -ef | grep ${devProxyExe} | grep -v grep | awk '{print $2}'"`);
         return processId.trim() !== '';
     };
+    if (platform === 'linux') {
+        const processId = await executeCommand(`/bin/bash -c "ps -ef | grep ${devProxyExe} | grep -v grep | awk '{print $2}'"`);
+        return processId.trim() !== '';
+    }
     return false;
 };